.gitattributes: Don't diff more files.
Qt config (Trolltech.conf) is generated, and we only care about one of
the properties so far.

Supertuxcart config may store other input device config we don't care to
acknowledge yet.

lib/profile is a plan 9 rc thing, but we want to ignore that most of the
time, if deleted, because it looks ugly on XDG systems' $HOME.