.vimrc: Remove gvim configuration
Am not going to use gvim. I've tried gnvim, but still not feeling the
GUI love.

As for NeoVim Qt, that's only a use case on Windows, to compensate for
Windows laundry list of issues and misdesigns, technical or UX. If I'm
stuck in a room with only a Windows laptop for weeks, fine, I might
set some things again, but that's not supposed to be a steady state.