.config/nvim: Replace slow/sync highlight/regex with tree-sitter
I'm on nightly (or rather, git repo) NeoVim 0.6-rcwhatever. Basic
tree-sitter support is in the repo, minus config.

Configure the bare minimum. Mappings still subject to change.

The molokai (not monokai) colouring of Python and vimrc-embedded-lua
look much better now, on par with the also improved but not as
drastically JS and already beautiful Ruby.

Also, because tree-sitter is finally async (or rather, less blocking,
still steals some time because it runs in the same thread's libuv), the
whole editor feels MUCH faster.